+++ Write_code_to_draw/Shapes.cs
@@ -19,7 +19,7 @@ namespace Write_code_to_draw
         /// <param name="x"></param> initial x point in the picturebox
         /// <param name="y"></param> initial y point in the picturebox
         /// <param name="fill"></param> fill on or off in the application
-        void Draw(Graphics G, int x, int y,bool fill);
+        void Draw(Graphics G, float x, float y,bool fill);
